test(runner): make guest download fixture deterministic
